execution: Rename package from task_execution to execution
Rename execution_interface to interface
