+++ setup.py
@@ -6,7 +6,7 @@ setup(name='mrs',
       packages=['mrs', 'mrs.config', 'mrs.config.builders', 'mrs.config.default',
                 'mrs.db.models', 'mrs.db.models.performance', 'mrs.exceptions',
                 'mrs.experiments', 'mrs.structs', 'mrs.allocation', 'mrs.bidding',
-                'mrs.scheduling', 'mrs.dispatching', 'mrs.task_execution', 'mrs.utils'],
+                'mrs.scheduling', 'mrs.dispatching', 'mrs.execution', 'mrs.utils'],
       version='0.2.0',
       description='Multi-Robot System (MRS) components for performing'
                   'Multi-Robot Task Allocation (MRTA) and executing'
